refactor: 수정 Home page 컴포넌트화 완료
-Home 페이지를 3개의 컴포넌트로 나누어 리팩토링 진행

-(진행중)type any로 지정해준 것 리팩토링 필요
-(해결)컴포넌트화 리팩토링 필요
-(진행중)유저 프로필사진 가져와서 등록해주는 부분 필요
-(신규)props로 내려준걸 Slice로 넣어줄 수 있는지 확인 및 진행 필요
