+++ .gitmessage.txt
@@ -1,13 +1,12 @@
 # <타입> : <제목> 형식으로 작성하며 제목은 최대 50글자 정도로만 입력
 # 제목을 아랫줄에 작성, 제목 끝에 마침표 금지, 무엇을 했는지 명확하게 작성
-feat: 추가 getPost, getCategory 완성
+feat: 추가 전체카테고리 완성
 ################
 # 본문(추가 설명)을 아랫줄에 작성
--Home(page) Post 불러오기 완성
--Category 별로 Post 불러오기 완성
+-데이터에 'all' 카테고리를 기본설정으로 하여 전체 눌렀을 때 글 전부 불러올 수 있게 기능 추가
 ################
 # 꼬릿말(footer)을 아랫줄에 작성 (관련된 이슈 번호 등 추가)
--전체보기 구성해야함
+-(해결)전체보기 구성해야함
 -type any로 지정해준 것 리팩토링 필요
 -컴포넌트화 리팩토링 필요
 -유저 프로필사진 가져와서 등록해주는 부분 필요
