+++ src/redux/modules/mainSlice.ts
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function mainSlice() {
+  
+}
