+++ src/shared/type.ts
@@ -1,0 +1,10 @@
+export interface PostState {
+    id: string;
+    nickname: string;
+    category: any;
+    comment: string[];
+    content: string;
+    createdAt: any;
+    title: string;
+    userid: number;
+  }
